refactor: move dispatch and autocomplete into CommandRegistry
Moves `parseSlashCommand`, `dispatchInput`, `matchCommands`,
`filterCommands`, `applyArguments`, `SlashCommandResult`, and
`DispatchResult` from `input.ts` into `command-registry.ts`, where the
registry is already the authoritative source of all command knowledge.

The registry now exposes:
- `registry.parseSlashCommand(input)` — parse a slash command token
- `registry.dispatch(input, readFile?)` — full REPL dispatch
- `registry.suggest(query, listDir?, readFile?)` — filtered autocomplete

`input.ts` is now focused on core input primitives: `ask()`, `pick()`,
multiline input, and stash handling.

Closes #676
