+++ CLAUDE.md
@@ -86,6 +86,6 @@ See **`docs/type-system.md`** for the full design. In brief: one server model cl
 - Pass model objects between internal methods, not raw ID strings. Controllers parse IDs from wire messages first, look up the model object (e.g. `Worker.get(workerId)`, `Task.get(taskId)`), then pass the object to helpers. `task.assign(worker: Worker)`, `sendMsg(worker: Worker, ...)`, `cancelWorker(worker, task)`, etc. all take objects. Use `import type` for circular type-only references (e.g. `task.ts` imports `Worker` type, `worker.ts` imports `Task` type).
 - In unit tests, use `setupInMemoryTasks()` from `tests/helpers/task.ts` instead of `initDb()` — it mocks the `Task` layer with an in-memory map. `WebhookEvent.log()` and `ForemanMessage.log()` return `Promise<void>`; production callers use `void` to discard the promise; tests can `await` them for deterministic setup. `tests/setup.ts` (a vitest `setupFiles` entry) always initialises the DB with the in-memory shim from `tests/helpers/memory-db.ts` so that these calls no-op cleanly without guards in production code; DB tests override this with `initDb(createTestSupabase())` at module level.
 - All wire types live in `shared/wire.ts` and are imported as `import * as Wire from "../../shared/wire.js"` (adjust relative path as needed). `src/wire.ts` re-exports from `shared/wire.ts` for backward compatibility. The DB model class for foreman messages is `ForemanMessage` (in `src/foreman/models/foreman-message.ts`), distinct from the wire type `Wire.ForemanMessage`.
-- In agent tests that need slash commands, create a `CommandRegistry` instance directly (`new CommandRegistry()`) and populate it via `registerTestCommands()` from `tests/helpers.ts` (which returns the registry). Pass the registry to functions like `dispatchInput`, `parseSlashCommand`, etc.; call `registry.listCommandNames()` and `registry.listCommands()` directly on the registry (with injectable `listDir`/`readFile` args for testing). When calling `registerWorkspaceCommands` or `registerWorkerCommands` in tests, pass a pre-scoped registry (e.g. `registry.scoped("workspace")`) — scoping is always the caller's responsibility. `registerWorkspaceCommands` takes `Workspace | undefined` as its first argument (`undefined` means no GitHub repo configured — commands degrade gracefully). `Workspace` is constructed with `(workspaceDir, sessionId, repoUrl, originalCwd, confirm)` at startup before any clone; `workspace.isCreated` tracks whether `create()` has been called; `WorkerSession.workspace` exposes the session's workspace for command registration.
+- In agent tests that need slash commands, create a `CommandRegistry` instance directly (`new CommandRegistry()`) and populate it via `registerTestCommands()` from `tests/helpers.ts` (which returns the registry). Use `registry.dispatch(input, readFile?)` for dispatch, `registry.parseSlashCommand(input)` for slash command parsing, `registry.suggest(query, listDir?, readFile?)` for autocomplete filtering; call `registry.listCommandNames()` and `registry.listCommands()` directly on the registry (with injectable `listDir`/`readFile` args for testing). `applyArguments`, `matchCommands`, `filterCommands`, `SlashCommandResult`, and `DispatchResult` are exported from `command-registry.ts`. When calling `registerWorkspaceCommands` or `registerWorkerCommands` in tests, pass a pre-scoped registry (e.g. `registry.scoped("workspace")`) — scoping is always the caller's responsibility. `registerWorkspaceCommands` takes `Workspace | undefined` as its first argument (`undefined` means no GitHub repo configured — commands degrade gracefully). `Workspace` is constructed with `(workspaceDir, sessionId, repoUrl, originalCwd, confirm)` at startup before any clone; `workspace.isCreated` tracks whether `create()` has been called; `WorkerSession.workspace` exposes the session's workspace for command registration.
 - In foreman tests that use `new ForemanWss(...)`, call `Worker._reset()` in `beforeEach` to clear the module-level worker registry between tests. `ForemanWss` takes a single options object `{ config, taskManager, server, adminWss? }` — `Worker` is imported directly by `wss.ts` (no registry injection). To unit-test hello handlers or event routing, instantiate `ForemanWss` with `http.createServer()`, a mock config, and a mock `taskManager` (include `on: vi.fn()` in the mock); then spy with `vi.spyOn(wss, "sendMsg").mockImplementation(() => {})` and `vi.spyOn(utils, "log").mockImplementation(() => {})` (import `* as utils from "../src/utils.js"`) before calling methods like `wss.handleBusyHello(...)`, `wss.handleIdleHello(...)`, `wss.routePrEvent(...)`, or `wss.forwardEvent(...)`. Mock `taskManager.assignIdleWorkers()` to return `[]` and `taskManager.fetchAndLoadDeps()` to return `Promise.resolve()`.
 - In browser tests, the server is shared across all tests — use unique issue numbers per test rather than resetting server state.
