+++ tests/repl.slash.test.ts
@@ -1,5 +1,4 @@
 import { describe, it, expect, beforeEach } from "vitest";
-import { parseSlashCommand } from "../src/agent/input.js";
 import { resolveCommandFilePath, resolveContent, type CommandRegistry } from "../src/agent/command-registry.js";
 import { registerTestCommands } from "./helpers.js";
 
@@ -9,66 +8,66 @@ beforeEach(async () => { registry = await registerTestCommands(); });
 
 describe("parseSlashCommand", () => {
   it("returns null for non-slash input", () => {
-    expect(parseSlashCommand("hello", registry)).toBeNull();
+    expect(registry.parseSlashCommand("hello")).toBeNull();
   });
 
   it("returns null for empty string", () => {
-    expect(parseSlashCommand("", registry)).toBeNull();
+    expect(registry.parseSlashCommand("")).toBeNull();
   });
 
   it("returns null for bare slash", () => {
-    expect(parseSlashCommand("/", registry)).toBeNull();
+    expect(registry.parseSlashCommand("/")).toBeNull();
   });
 
   it("recognizes /exit as builtin", () => {
-    expect(parseSlashCommand("/exit", registry)).toEqual({ type: "command", name: "exit" });
+    expect(registry.parseSlashCommand("/exit")).toEqual({ type: "command", name: "exit" });
   });
 
   it("recognizes /clear as builtin", () => {
-    expect(parseSlashCommand("/clear", registry)).toEqual({ type: "command", name: "clear" });
+    expect(registry.parseSlashCommand("/clear")).toEqual({ type: "command", name: "clear" });
   });
 
   it("returns unknown for unrecognized command with no file", () => {
-    const result = parseSlashCommand("/unknown", registry);
+    const result = registry.parseSlashCommand("/unknown");
     expect(result).toEqual({ type: "unknown_command", command: "unknown" });
   });
 
   it("parses command name from input with arguments", () => {
-    const result = parseSlashCommand("/foo some args", registry);
+    const result = registry.parseSlashCommand("/foo some args");
     expect(result).toEqual({ type: "unknown_command", command: "foo" });
   });
 
   it("parses command name with colon namespace", () => {
-    const result = parseSlashCommand("/foo:bar", registry);
+    const result = registry.parseSlashCommand("/foo:bar");
     expect(result).toEqual({ type: "unknown_command", command: "foo:bar" });
   });
 
   it("recognizes /worker:complete (canonical name)", () => {
-    expect(parseSlashCommand("/worker:complete", registry)).toEqual({ type: "command", name: "worker:complete" });
+    expect(registry.parseSlashCommand("/worker:complete")).toEqual({ type: "command", name: "worker:complete" });
   });
 
   it("recognizes /workspace:create (canonical name)", () => {
-    expect(parseSlashCommand("/workspace:create", registry)).toEqual({ type: "command", name: "workspace:create" });
+    expect(registry.parseSlashCommand("/workspace:create")).toEqual({ type: "command", name: "workspace:create" });
   });
 
   it("recognizes /workspace:reset (canonical name)", () => {
-    expect(parseSlashCommand("/workspace:reset", registry)).toEqual({ type: "command", name: "workspace:reset" });
+    expect(registry.parseSlashCommand("/workspace:reset")).toEqual({ type: "command", name: "workspace:reset" });
   });
 
   it("recognizes /workspace:remove (canonical name)", () => {
-    expect(parseSlashCommand("/workspace:remove", registry)).toEqual({ type: "command", name: "workspace:remove" });
+    expect(registry.parseSlashCommand("/workspace:remove")).toEqual({ type: "command", name: "workspace:remove" });
   });
 
   it("recognizes /workspace:prune (canonical name)", () => {
-    expect(parseSlashCommand("/workspace:prune", registry)).toEqual({ type: "command", name: "workspace:prune" });
+    expect(registry.parseSlashCommand("/workspace:prune")).toEqual({ type: "command", name: "workspace:prune" });
   });
 
   it("recognizes /model", () => {
-    expect(parseSlashCommand("/model", registry)).toEqual({ type: "command", name: "model" });
+    expect(registry.parseSlashCommand("/model")).toEqual({ type: "command", name: "model" });
   });
 
   it("recognizes /effort", () => {
-    expect(parseSlashCommand("/effort", registry)).toEqual({ type: "command", name: "effort" });
+    expect(registry.parseSlashCommand("/effort")).toEqual({ type: "command", name: "effort" });
   });
 });
